(1) Changed rint to ROUND because of cygwin problems. (2) Variable new_val win Pve5.c was not correctly allocated.
git-svn-id: svn://141.35.69.214/cat12@78 9f44b166-36c4-483c-8295-4af200601734

+++ Amap.h
@@ -39,7 +39,7 @@
 #endif
 
 #ifndef ROUND
-#define ROUND( x ) ((int) ((x) + ( ((x) >= 0) ? 0.5 : (-0.5) ) ))
+#define ROUND( x ) ((long) ((x) + ( ((x) >= 0) ? 0.5 : (-0.5) ) ))
 #endif
 
 struct point {
